+++ src/oci/database/models/create_autonomous_database_base.py
@@ -196,6 +196,10 @@ class CreateAutonomousDatabaseBase(object):
             The value to assign to the autonomous_container_database_id property of this CreateAutonomousDatabaseBase.
         :type autonomous_container_database_id: str
 
+        :param in_memory_percentage:
+            The value to assign to the in_memory_percentage property of this CreateAutonomousDatabaseBase.
+        :type in_memory_percentage: int
+
         :param is_access_control_enabled:
             The value to assign to the is_access_control_enabled property of this CreateAutonomousDatabaseBase.
         :type is_access_control_enabled: bool
@@ -318,6 +322,7 @@ class CreateAutonomousDatabaseBase(object):
             'is_auto_scaling_enabled': 'bool',
             'is_dedicated': 'bool',
             'autonomous_container_database_id': 'str',
+            'in_memory_percentage': 'int',
             'is_access_control_enabled': 'bool',
             'whitelisted_ips': 'list[str]',
             'are_primary_whitelisted_ips_used': 'bool',
@@ -367,6 +372,7 @@ class CreateAutonomousDatabaseBase(object):
             'is_auto_scaling_enabled': 'isAutoScalingEnabled',
             'is_dedicated': 'isDedicated',
             'autonomous_container_database_id': 'autonomousContainerDatabaseId',
+            'in_memory_percentage': 'inMemoryPercentage',
             'is_access_control_enabled': 'isAccessControlEnabled',
             'whitelisted_ips': 'whitelistedIps',
             'are_primary_whitelisted_ips_used': 'arePrimaryWhitelistedIpsUsed',
@@ -415,6 +421,7 @@ class CreateAutonomousDatabaseBase(object):
         self._is_auto_scaling_enabled = None
         self._is_dedicated = None
         self._autonomous_container_database_id = None
+        self._in_memory_percentage = None
         self._is_access_control_enabled = None
         self._whitelisted_ips = None
         self._are_primary_whitelisted_ips_used = None
@@ -1144,6 +1151,30 @@ class CreateAutonomousDatabaseBase(object):
         :type: str
         """
         self._autonomous_container_database_id = autonomous_container_database_id
+
+    @property
+    def in_memory_percentage(self):
+        """
+        Gets the in_memory_percentage of this CreateAutonomousDatabaseBase.
+        The percentage of the System Global Area(SGA) assigned to In-Memory tables in Autonomous Database.
+
+
+        :return: The in_memory_percentage of this CreateAutonomousDatabaseBase.
+        :rtype: int
+        """
+        return self._in_memory_percentage
+
+    @in_memory_percentage.setter
+    def in_memory_percentage(self, in_memory_percentage):
+        """
+        Sets the in_memory_percentage of this CreateAutonomousDatabaseBase.
+        The percentage of the System Global Area(SGA) assigned to In-Memory tables in Autonomous Database.
+
+
+        :param in_memory_percentage: The in_memory_percentage of this CreateAutonomousDatabaseBase.
+        :type: int
+        """
+        self._in_memory_percentage = in_memory_percentage
 
     @property
     def is_access_control_enabled(self):