+++ CHANGELOG.rst
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`_.
 ====================
+2.110.2 - 2023-08-22
+====================
+
+Added
+-----
+* Support for Compute Cloud at Customer service
+* Support for warehouse data objects in the Operations Insights service
+* Support for standard queries on operations Insights data objects in the Operations Insights service
+* Support for database in memory on autonomous database create operations in the Database service
+ 
+====================
 2.110.1 - 2023-08-15
 ====================